requirements.txt
astroid==2.1.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mss==4.0.1
Pillow==5.4.1
pkg-resources==0.0.0
pylint==2.2.2
PyQt5==5.9.2
sip==4.19.8
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
